+++ tests/API/api.cmake
@@ -2,7 +2,7 @@
 macro(api_test_executable NAME DIRECTORIES)
   sos_sdk_app_target(RELEASE ${NAME} "unittest" release ${SOS_ARCH})
   add_executable(${RELEASE_TARGET})
-  add_dependencies(api_tests ${RELEASE_TARGET})
+  add_dependencies(API_test ${RELEASE_TARGET})
 
   target_sources(${RELEASE_TARGET}
     PRIVATE
@@ -55,6 +55,7 @@ macro(api_test_executable NAME DIRECTORIES)
 
   if (SOS_IS_LINK)
     sos_sdk_add_test(${NAME} "unittest" release)
+    set_tests_properties(${NAME}_release PROPERTIES DEPENDS API_Test)
     #sos_sdk_add_test(${NAME} "unittest" coverage)
   endif ()